# apps/server/api/openapi.yaml
openapi: "3.1.0"
info:
  title: LIRS API
  version: "0.1.0"
servers:
  - url: http://localhost:8081
paths:
...
  /api/procurement-projects:
    get:
      summary: List procurement project names and numbers for purchasable materials
      responses:
        "200":
          description: Procurement project list
    post:
      summary: Create or update a procurement project by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcurementProjectInput"
      responses:
        "200":
          description: Procurement project saved
  /api/procurement-projects/{id}:
    patch:
      summary: Update a procurement project, including expiry and status
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcurementProjectInput"
      responses:
        "200":
          description: Procurement project updated
    delete:
      summary: Disable a procurement project
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Procurement project disabled
  /api/purchasable-materials:
    get:
      summary: List purchasable material catalog with procurement project status
      responses:
        "200":
          description: Purchasable material list
    post:
      summary: Create or update a purchasable material and auto-create missing procurement project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchasableMaterialInput"
      responses:
        "200":
          description: Purchasable material saved
  /api/purchasable-materials/{id}:
    patch:
      summary: Update a purchasable material and auto-create missing procurement project
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchasableMaterialInput"
      responses:
        "200":
          description: Purchasable material updated
    delete:
      summary: Delete a purchasable material
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Purchasable material deleted
...
components:
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ProcurementProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        expiresAt: { type: string, format: date }
        status:
          type: string
          enum: [active, disabled]
    PurchasableMaterialInput:
      type: object
      required: [idNo, sequenceNo, projectName, brand, spec, unit, purchasePrice]
      properties:
        idNo: { type: string }
        sequenceNo: { type: string }
        procurementProjectId: { type: string, format: uuid }
        procurementProject: { type: string }
        projectName: { type: string }
        brand: { type: string }
        spec: { type: string }
        unit: { type: string }
        purchasePrice: { type: number }
        remark: { type: string }
        technicalRequirement: { type: string }
        minSpec: { type: string }
...